+++ tickets.php
@@ -434,9 +434,11 @@ $user_active_assigned_tickets = $row['total_tickets_assigned'];
                                       <div class="dropdown-menu">
                                           <a class="dropdown-item" href="#" data-toggle="modal"
                                              data-target="#editTicketModal<?php echo $ticket_id; ?>">Edit</a>
-                                          <div class="dropdown-divider"></div>
-                                          <a class="dropdown-item text-danger"
-                                             href="post.php?delete_ticket=<?php echo $ticket_id; ?>">Delete</a>
+                                        <?php if($session_user_role == 3) { ?>
+                                            <div class="dropdown-divider"></div>
+                                            <a class="dropdown-item text-danger"
+                                               href="post.php?delete_ticket=<?php echo $ticket_id; ?>">Delete</a>
+                                        <?php } ?>
                                       </div>
                                   </div>
                                 <?php } ?>
